boards: rpi_pico2: Create CPU1 board variant
The rp2350 has 2 processor cores that can be either Cortex-M33 or
Hazard3 riscv architecture. Add a board variant to the rpi_pico2 for the
Cortex-M33 CPU1 that runs from a reserved 256KB block in SRAM and has
the minimum devices allocated to it.

The rpi_pico2/rp2350a/m33/cpu1 board target has neither a UART console,
access to the flash, nor the default enabled peripherals to avoid
conflicts with the main CPU. However, CPU1 does enable the &gpio0
devicetree node since it has dedicated access through the SIO block. It
is the application writer's responsibility to divvy up the GPIOs between
cores as they need.

Co-authored-by: Dmitrii Sharshakov <d3dx12.xx@gmail.com>
Signed-off-by: Dan Collins <dan@collinsnz.com>
Signed-off-by: Dmitrii Sharshakov <d3dx12.xx@gmail.com>
Signed-off-by: Andrew Fantino <fantino@qti.qualcomm.com>